fix(macos): extract SidebarView to break MainWindowView body cascade (LUM-1082)
Monolithic MainWindowView.body evaluated the entire view hierarchy as one
unit, so SidebarConversationsHeader (a leaf in the sidebar path) was
re-evaluated on every state change anywhere in the window — triggering
2s main-thread hangs (Sentry MACOS-K2).

Move the sidebar into a standalone `SidebarView: View` that takes only the
dependencies it reads, so its body re-evaluates only when sidebar-relevant
state changes. Same shape as the LUM-655 extraction work.

Also make SidebarConversationsHeader Equatable so .equatable() can
short-circuit re-evaluation when its value props are unchanged.

Pure refactor — no behavior change, narrower invalidation scope.
