+++ clients/macos/vellum-assistant/Features/MainWindow/Sidebar/SidebarConversationsHeader.swift
@@ -1,7 +1,7 @@
 import SwiftUI
 import VellumAssistantShared
 
-struct SidebarConversationsHeader: View {
+struct SidebarConversationsHeader: View, Equatable {
     let hasUnseenConversations: Bool
     var isLoading: Bool = false
     let onMarkAllSeen: () -> Void
@@ -9,6 +9,15 @@ struct SidebarConversationsHeader: View {
     var onCreateGroup: (() -> Void)? = nil
 
     @AppStorage("newChatShortcut") private var newChatShortcut: String = "cmd+n"
+
+    /// Closures are assumed stable across renders — only compare the value
+    /// props that actually drive rendering. Lets `.equatable()` short-circuit
+    /// body evaluation when parent invalidates without relevant changes.
+    static func == (lhs: SidebarConversationsHeader, rhs: SidebarConversationsHeader) -> Bool {
+        lhs.hasUnseenConversations == rhs.hasUnseenConversations
+            && lhs.isLoading == rhs.isLoading
+            && (lhs.onCreateGroup == nil) == (rhs.onCreateGroup == nil)
+    }
 
     private var newChatTooltip: String {
         let label = "New conversation"
